add doc strings, moved implementation details, add plannerStatus as output
