+++ src/planner/vectorfield/detail/VectorFieldPlannerExceptions.cpp
@@ -1,4 +1,4 @@
-#include <aikido/planner/vectorfield/VectorFieldPlannerExceptions.hpp>
+#include <aikido/planner/vectorfield/detail/VectorFieldPlannerExceptions.hpp>
 
 namespace aikido {
 namespace planner {
