+++ tests/planner/vectorfield/test_VectorFieldPlanner.cpp
@@ -5,6 +5,8 @@
 #include <aikido/constraint/Testable.hpp>
 #include <aikido/distance/defaults.hpp>
 #include <aikido/planner/PlanningResult.hpp>
+#include <aikido/planner/PlanningResult.hpp>
+#include <aikido/planner/vectorfield/MoveEndEffectorOffsetVectorField.hpp>
 #include <aikido/planner/vectorfield/VectorFieldPlanner.hpp>
 #include <aikido/planner/vectorfield/VectorFieldUtil.hpp>
 #include <aikido/statespace/GeodesicInterpolator.hpp>
@@ -337,8 +339,9 @@ TEST_F(VectorFieldPlannerTest, PlanToEndEffectorOffsetTest)
   double initialStepSize = 0.05;
   double jointLimitTolerance = 1e-3;
   double constraintCheckResolution = 1e-3;
-  double timelimit = 5.;
+  std::chrono::duration<double> timelimit(5.);
   double maxDistance = MoveEndEffectorOffsetVectorField::InvalidMaxDistance;
+  aikido::planner::PlanningResult planningResult;
 
   auto traj = aikido::planner::vectorfield::planToEndEffectorOffset(
       mStateSpace,
@@ -353,7 +356,8 @@ TEST_F(VectorFieldPlannerTest, PlanToEndEffectorOffsetTest)
       initialStepSize,
       jointLimitTolerance,
       constraintCheckResolution,
-      timelimit);
+      timelimit,
+      planningResult);
 
   EXPECT_FALSE(traj == nullptr) << "Trajectory not found";
 
@@ -419,8 +423,9 @@ TEST_F(VectorFieldPlannerTest, DirectionZeroVector)
   double initialStepSize = 0.05;
   double jointLimitTolerance = 1e-3;
   double constraintCheckResolution = 1e-3;
-  double timelimit = 5.;
+  std::chrono::duration<double> timelimit(5.);
   double maxDistance = MoveEndEffectorOffsetVectorField::InvalidMaxDistance;
+  aikido::planner::PlanningResult planningResult;
 
   mStateSpace->getMetaSkeleton()->setPositions(mStartConfig);
 
@@ -438,7 +443,8 @@ TEST_F(VectorFieldPlannerTest, DirectionZeroVector)
           initialStepSize,
           jointLimitTolerance,
           constraintCheckResolution,
-          timelimit),
+          timelimit,
+          planningResult),
       std::runtime_error);
 }
 
@@ -460,7 +466,8 @@ TEST_F(VectorFieldPlannerTest, PlanToEndEffectorPoseTest)
   double initialStepSize = 0.05;
   double jointLimitTolerance = 1e-3;
   double constraintCheckResolution = 1e-3;
-  double timelimit = 5.;
+  std::chrono::duration<double> timelimit(5.);
+  aikido::planner::PlanningResult planningResult;
 
   mStateSpace->getMetaSkeleton()->setPositions(mStartConfig);
 
@@ -475,7 +482,8 @@ TEST_F(VectorFieldPlannerTest, PlanToEndEffectorPoseTest)
       initialStepSize,
       jointLimitTolerance,
       constraintCheckResolution,
-      timelimit);
+      timelimit,
+      planningResult);
 
   EXPECT_FALSE(traj1 == nullptr) << "Trajectory not found";
